requirements.txt
flask
requests
beautifulsoup4
flask-cors
gunicorn
psycopg2-binary